Measure runtime accurately
Currently, runtime is assumed to be same as user specified time.
Nevertheless, the time can be different with real runtime because memory
walking function check the ending signal with an interval.  Because the
interval is in proportion to walking memory size, the error can be large
as walking memory size grows.  This commit fixes the problem by
accurately measuring the start time and end time of the workload and
then using that to calculate the runtime.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>