Set ending signal check interval based on accesses count
To minimize ending signal checking noise, the logic check ending signal
once for hundred iteration of memory walk.  As a consequence, the
interval grows as memory size grows.  For couple of tens of gibibytes
memory, the program runs very long time though user specifies the
runtime to be short.  To solve the problem, this commit modifies the
ending signal check logic to set the interval as loughly one billion
memory accesses.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>